+++ src/Component/ContentElement/GridWrapperElementController.php
@@ -29,10 +29,16 @@ final class GridWrapperElementController extends AbstractContentElementControlle
     #[Override]
     protected function getResponse(FragmentTemplate $template, ContentModel $model, Request $request): Response
     {
-        $template->iterator  = $this->getIterator($model);
-        $template->name      = $model->bs_grid_name;
-        $template->color     = $this->colorRotate->getColor('ce:' . $model->id);
-        $template->isBackend = $this->isBackendScope($request);
+        if ($this->isBackendScope($request)) {
+            $template->setName('backend/grid_wildcard');
+
+            $template->set('title', $model->bs_grid_name);
+            $template->set('color', $this->colorRotate->getColor('ce:' . $model->id));
+
+            return $template->getResponse();
+        }
+
+        $template->set('iterator', $this->getIterator($model));
 
         return $template->getResponse();
     }