+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Changelog
 [Unreleased]
 ------------
 
+3.1.0 (2026-06-05)
+------------------
+
 ### Deprecated
 
  - Legacy content elements `bs_gridStart`, `bs_gridStop` and `bs_gridSeparator` are deprecated.